Update core.js
Update to the "core" of core to include a Navigator Object. openScreen function, and all subsequent navigation would be part of the core object, so that its more easily overridden when using widgets or other custom navigation patterns.

Additionally added a `_modal` flag to the openScreen function to account for modal windows within the application.